SNN_ECG_Reproducibility_Review_Package_v2/scripts/eval_snn.py
import argparse, json, yaml, pathlib, numpy as np

def main():
    ap = argparse.ArgumentParser()
    ap.add_argument("--config", type=str, required=True)
    ap.add_argument("--splits", type=str, required=True)
    ap.add_argument("--data", type=str, required=True)
    ap.add_argument("--ckpt", type=str, required=True)
    args = ap.parse_args()

    cfg = yaml.safe_load(open(args.config, "r", encoding="utf-8"))
    splits = json.load(open(args.splits, "r", encoding="utf-8"))
    data_dir = pathlib.Path(args.data)

    print("[INFO] Evaluation SNN stub.")
    print("[CKPT]", args.ckpt)
    print("[CFG ]", cfg["reporting"]["metrics"] if "reporting" in cfg else "(no metrics)")
    print("[DATA] Expecting NumPy arrays in:", data_dir.as_posix())
    print("[NOTE] Replace this with your evaluation function to compute accuracy/precision/recall/F1/AUROC/AUPRC.")

if __name__ == "__main__":
    main()